src/app/api/ptl-tools/requirements.txt
pdfplumber==0.11.3
python-docx==1.1.2